Create a SQL database
This code is a supplementary program to the webscrape program that puts the text files into a SQLite database. For ogranizational purposes, this program is separate from webscrape.py, but could be merged with it. In that case the link_list variable could be replaced with variable "finalized" from webscrape.py